Introspect during construction
