+++ src/main/java/com/fasterxml/jackson/databind/jsontype/impl/AsDeductionTypeDeserializer.java
@@ -41,9 +41,9 @@ public class AsDeductionTypeDeserializer extends AsPropertyTypeDeserializer
     private final Map<BitSet, String> subtypeFingerprints;
 
     public AsDeductionTypeDeserializer(JavaType bt, TypeIdResolver idRes, JavaType defaultImpl,
-            DeserializationConfig config, Collection<NamedType> subtypes)
+            DeserializationConfig config, Collection<NamedType> subtypes, boolean hasTypeResolverAnnotation)
     {
-        super(bt, idRes, null, false, defaultImpl, null);
+        super(bt, idRes, null, false, defaultImpl, null, hasTypeResolverAnnotation);
         fieldBitIndex = new HashMap<>();
         subtypeFingerprints = buildFingerprints(config, subtypes);
     }
