+++ src/main/java/com/fasterxml/jackson/databind/jsontype/impl/StdTypeResolverBuilder.java
@@ -4,12 +4,11 @@ import java.util.Collection;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
-import com.fasterxml.jackson.databind.DeserializationConfig;
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.MapperFeature;
-import com.fasterxml.jackson.databind.SerializationConfig;
+import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.annotation.NoClass;
 import com.fasterxml.jackson.databind.cfg.MapperConfig;
+import com.fasterxml.jackson.databind.introspect.AnnotatedClass;
+import com.fasterxml.jackson.databind.introspect.AnnotatedClassResolver;
 import com.fasterxml.jackson.databind.jsontype.*;
 import com.fasterxml.jackson.databind.util.ClassUtil;
 
@@ -162,7 +161,7 @@ public class StdTypeResolverBuilder
 
         if(_idType == JsonTypeInfo.Id.DEDUCTION) {
             // Deduction doesn't require an includeAs property
-            return new AsDeductionTypeDeserializer(baseType, idRes, defaultImpl, config, subtypes);
+            return new AsDeductionTypeDeserializer(baseType, idRes, defaultImpl, config, subtypes, _hasTypeResolverAnnotations(config, baseType));
         }
 
         // First, method for converting type info to type id:
@@ -173,7 +172,7 @@ public class StdTypeResolverBuilder
         case PROPERTY:
         case EXISTING_PROPERTY: // as per [#528] same class as PROPERTY
             return new AsPropertyTypeDeserializer(baseType, idRes,
-                    _typeProperty, _typeIdVisible, defaultImpl, _includeAs);
+                    _typeProperty, _typeIdVisible, defaultImpl, _includeAs, _hasTypeResolverAnnotations(config, baseType));
         case WRAPPER_OBJECT:
             return new AsWrapperTypeDeserializer(baseType, idRes,
                     _typeProperty, _typeIdVisible, defaultImpl);
@@ -182,6 +181,22 @@ public class StdTypeResolverBuilder
                     _typeProperty, _typeIdVisible, defaultImpl);
         }
         throw new IllegalStateException("Do not know how to construct standard type serializer for inclusion type: "+_includeAs);
+    }
+
+    /**
+     * Checks whether the given class has annotations indicating some type resolver
+     * is applied, for example {@link com.fasterxml.jackson.annotation.JsonSubTypes}.
+     * Only initializes {@link #_hasTypeResolverAnnotations} once if its value is null.
+     *
+     * @param config the deserialization configuration to use
+     * @param baseType the base type to check for type resolver annotations
+     * @return true if the class has type resolver annotations, false otherwise
+     * @since 2.15
+     */
+    protected boolean _hasTypeResolverAnnotations(DeserializationConfig config, JavaType baseType){
+        AnnotatedClass ac = AnnotatedClassResolver.resolveWithoutSuperTypes(config,  baseType.getRawClass());
+        AnnotationIntrospector ai = config.getAnnotationIntrospector();
+        return ai.findTypeResolver(config, ac, baseType) != null;
     }
 
     protected JavaType defineDefaultImpl(DeserializationConfig config, JavaType baseType)
